requirements.txt
alembic==1.12.0
anaconda-anon-usage @ file:///C:/b/abs_c3w_h1zzjg/croot/anaconda-anon-usage_1710965204622/work
annotated-types==0.7.0
anyio==3.7.1
archspec @ file:///croot/archspec_1709217642129/work
auto_subtitle @ git+https://github.com/m1guelpf/auto-subtitle.git@124ccb1ac17d5b7a27dd81b3d8f8fed6ef1a5408
bcrypt==4.0.1
boltons @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/boltons_1699480450092/work
Brotli @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/brotli-split_1699473013692/work
certifi @ file:///C:/b/abs_35d7n66oz9/croot/certifi_1707229248467/work/certifi
cffi @ file:///C:/b/abs_924gv1kxzj/croot/cffi_1700254355075/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
colorama @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/colorama_1699472650914/work
conda @ file:///C:/b/abs_1e6dlkntna/croot/conda_1710772093015/work
conda-content-trust @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/conda-content-trust_1699553484152/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1706733287605/work/src
conda-package-handling @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/conda-package-handling_1699480603217/work
conda_package_streaming @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/conda-package-streaming_1699475879769/work
cryptography @ file:///C:/b/abs_f5n93r0tun/croot/cryptography_1710350404202/work
distro @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/distro_1701796812765/work
ecdsa==0.19.0
fastapi==0.103.1
faiss-cpu==1.7.4
filelock==3.15.4
fsspec==2024.6.1
greenlet==3.0.3
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/idna_1699473483982/work
intel-openmp==2021.4.0
Jinja2==3.1.4
jsonpatch @ file:///C:/b/abs_d3zr1enxou/croot/jsonpatch_1710807549298/work
jsonpointer==2.1
libmambapy @ file:///C:/b/abs_7dmjutgtwb/croot/mamba-split_1712091963973/work/libmambapy
llvmlite==0.43.0
Mako==1.3.5
MarkupSafe==2.1.5
menuinst @ file:///C:/b/abs_099kybla52/croot/menuinst_1706732987063/work
mkl==2021.4.0
more-itertools==10.3.0
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==2.0.0
openai-whisper==20231117
packaging @ file:///C:/b/abs_cc1h2xfosn/croot/packaging_1710807447479/work
passlib==1.7.4
platformdirs @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/platformdirs_1701797392447/work
playwright==1.44.0
pluggy @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pluggy_1699472504117/work
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycosat @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pycosat_1699482932804/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
pyee==11.1.0
PySocks @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pysocks_1699473336188/work
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
regex==2024.5.15
requests @ file:///C:/b/abs_474vaa3x9e/croot/requests_1707355619957/work
rsa==4.9
ruamel.yaml @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/ruamel.yaml_1699483184324/work
sentence-transformers==2.2.2
setuptools==70.1.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.21
starlette==0.27.0
sympy==1.12.1
tbb==2021.13.0
tiktoken==0.7.0
torch==2.3.1
tqdm @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/tqdm_1701808178601/work
transformers==4.30.2
truststore @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/truststore_1701881385424/work
typing_extensions==4.12.0
urllib3 @ file:///C:/b/abs_4etpfrkumr/croot/urllib3_1707770616184/work
uvicorn==0.23.2
wheel==0.41.2
win-inet-pton @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/win_inet_pton_1699472992992/work
zstandard==0.19.0
